refactor: move base/topic.py to topics/models.py
