+++ workers/ohsome_quality_analyst/indicators/attribute_completeness/indicator.py.orig
@@ -7,7 +7,12 @@ import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
 from geojson import Feature
 
+<<<<<<< HEAD
+from ohsome_quality_analyst.base.topic import BaseTopic as Topic
 from ohsome_quality_analyst.indicators.base import BaseIndicator
+=======
+from ohsome_quality_analyst.base.indicator import BaseIndicator
+>>>>>>> fd093bd8 (refactor: move base/topic.py to topics/models.py)
 from ohsome_quality_analyst.ohsome import client as ohsome_client
 from ohsome_quality_analyst.topics.models import BaseTopic as Topic
