+++ workers/ohsome_quality_analyst/indicators/base.py
@@ -10,12 +10,12 @@ from typing import Dict, Literal, Optional
 import matplotlib.pyplot as plt
 from geojson import Feature
 
-from ohsome_quality_analyst.base.topic import BaseTopic as Topic
 from ohsome_quality_analyst.definitions import get_attribution, get_metadata
 from ohsome_quality_analyst.html_templates.template import (
     get_template,
     get_traffic_light,
 )
+from ohsome_quality_analyst.topics.models import BaseTopic as Topic
 from ohsome_quality_analyst.utils.helper import flatten_dict, json_serialize
