+++ workers/ohsome_quality_analyst/indicators/minimal/indicator.py
@@ -4,9 +4,9 @@ from string import Template
 import dateutil.parser
 from geojson import Feature
 
-from ohsome_quality_analyst.base.topic import BaseTopic as Topic
 from ohsome_quality_analyst.indicators.base import BaseIndicator
 from ohsome_quality_analyst.ohsome import client as ohsome_client
+from ohsome_quality_analyst.topics.models import BaseTopic as Topic
 
 
 class Minimal(BaseIndicator):
