+++ workers/tests/integrationtests/utils.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock
 import geojson
 import vcr
 
-from ohsome_quality_analyst.base.topic import TopicDefinition
 from ohsome_quality_analyst.definitions import get_topic_definition
+from ohsome_quality_analyst.topics.models import TopicDefinition
 
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 FIXTURE_DIR = os.path.join(TEST_DIR, "fixtures", "vcr_cassettes")
