+++ workers/tests/unittests/utils.py
@@ -2,8 +2,8 @@ import os
 
 import geojson
 
-from ohsome_quality_analyst.base.topic import TopicDefinition
 from ohsome_quality_analyst.definitions import get_topic_definition
+from ohsome_quality_analyst.topics.models import TopicDefinition
 
 
 def get_geojson_fixture(name):
